(#64) Move timer bars code to their own folder and files
Their namespace is kept the same for backwards compatibility.
