+++ Source/RAGENativeUI.csproj
@@ -84,7 +84,10 @@
     <Compile Include="Elements\Rectangle.cs" />
     <Compile Include="Elements\Text.cs" />
     <Compile Include="Elements\IElement.cs" />
-    <Compile Include="Elements\TimerBars.cs" />
+    <Compile Include="Elements\TimerBars\BarTimerBar.cs" />
+    <Compile Include="Elements\TimerBars\TextTimerBar.cs" />
+    <Compile Include="Elements\TimerBars\TimerBarBase.cs" />
+    <Compile Include="Elements\TimerBars\TimerBarPool.cs" />
     <Compile Include="Elements\UIMenuColoredItem.cs" />
     <Compile Include="Elements\InstructionalButtons.cs" />
     <Compile Include="Elements\UIMenuSwitchMenusItem.cs" />
